advanced_ai_agents/multi_agent_apps/agent_teams/multimodal_coding_agent_team/requirements_crewai.txt
crewai
crewai-tools
python-dotenv
openai
google-generativeai
streamlit
e2b-code-interpreter
Pillow
